Separated rendering of app and app files for testing.
